refactor: add student/teacher report class
